Wire up renderer tests to build system. CI tests working in Linux, OSX and Windows. r=vporof
